allow specifying unit for onboarding and uint64 for sizes

Closes #828

See merge request nunet/device-management-service!928

+++ dms/hardware/disk.go
@@ -34,7 +34,7 @@ func GetDisk() (types.Disk, error) {
 	}
 
 	return types.Disk{
-		Size: float64(totalStorage),
+		Size: totalStorage,
 	}, nil
 }
 
@@ -55,6 +55,6 @@ func GetDiskUsage() (types.Disk, error) {
 	}
 
 	return types.Disk{
-		Size: float64(usedStorage),
+		Size: usedStorage,
 	}, nil
 }
